refactor: DF-68 - Setup base class for PreviewPageController
