+++ designer/server/src/views/forms/editor-v2/guidance.njk
@@ -23,7 +23,9 @@
 
     <div class="govuk-grid-column-one-half-from-desktop">
       {%- include "../views/forms/editor-v2/partials/guidance-left-panel.njk" -%}
-
+    </div>
+    <div class="govuk-grid-column-one-half-from-desktop preview-panel page-preview-panel" id="preview-panel">
+      {%- include "../views/forms/editor-v2/partials/questions-right-panel.njk" -%}
     </div>
 
   {% endcall %}
