+++ model/src/form/form-editor/preview/controller/guidance-page-controller.js
@@ -1,0 +1,3 @@
+import { PreviewPageControllerBase } from '~/src/form/form-editor/preview/controller/page-controller-base.js'
+
+export class GuidancePageController extends PreviewPageControllerBase {}
