+++ model/src/form/form-editor/preview/index.js
@@ -21,3 +21,4 @@ export * from '~/src/form/form-editor/preview/long-answer.js'
 export * from '~/src/form/form-editor/preview/uk-address.js'
 export * from '~/src/form/form-editor/preview/yes-no.js'
 export * from '~/src/form/form-editor/preview/controller/page-controller.js'
+export { PagePreviewElements } from '~/src/form/form-editor/preview/controller/page-controller-base.js'
